// src/database/models/kysely-types.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Students {
    birthdate: Timestamp;
    cpf: string;
    email: string;
    gender: string;
    id?: string;
    name: string;
    phone: string;
    registered: boolean;
    subject: string;
    subjectId: number | null;
    teacherCpf: string | null;
}

export interface Subjects {
    description: string;
    hours: Numeric;
    id: Generated<number>;
    subjectName: string;
}

export interface Teachers {
    birthdate: Timestamp;
    cpf: string;
    email: string;
    gender: string;
    id?: string;
    name: string;
    phone: string;
    subject: string;
    subjectId: number | null;
}

export interface DB {
    students: Students;
    subjects: Subjects;
    teachers: Teachers;
}